move draw logic to separate package
