+++ packages/webaudio/package.json
@@ -34,6 +34,7 @@
   "homepage": "https://github.com/tidalcycles/strudel#readme",
   "dependencies": {
     "@strudel/core": "workspace:*",
+    "@strudel/canvas": "workspace:*",
     "superdough": "workspace:*"
   },
   "devDependencies": {
